requirements.txt
ruamel.yaml>=0.16.12
javaproperties>=0.8.0
mcdreforged>=2.1.0
dnspython>=2.1.0
mcstatus>=6.5.0
parse>=1.18.0
